Final 3D Metric
As the model can be trained on 2D slices, 3D volumes, or by using the sliding window technique, there is sometimes a need to test it on real 3D volumes, regardless of the training method used. Therefore, in this branch, we have introduced a 3D test function and added a parameter called 'test_inference' within the training function. This allows us to use it whenever needed, either before training if the model has already been pre-trained, or when the training finishes.